[WCD-CWS] wpa_supplicant: update current BSS level when signal change event occurs
When an “EVENT_SIGNAL_CHANGE” occurs the bgscan is informed about this change
but the new RSSI value is not stored. In consequence, when roaming candidates
are evaluated, the RSSI value of the current BSS used to compare is an old one
obtained during the last scan rather than the new one given by the signal change
event. This leads sometimes to bad decision when selecting a new BSS for
roaming.

This patch solves the issue by updating the current BSS level when receiving a
signal change event in order to have a very up-to-date current signal value when
choosing an new BSS.

---

BZ: 86937
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Orig-Change-Id: Id7346a9ea418b2f2d754ca1576d19f054e44163f

Change-Id: I99e8e4a9d2baaf180894893f60305c5c5e924f84
Signed-off-by: Matthieu Mauger <matthieux.mauger@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/40155
Reviewed-by: Alexander Bondar <alexander.bondar@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Patrick Combes <patrick.combes@intel.com>
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/40328
Tested-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: bae2db6141b7ffd15880f750f628aee774ae0126