P2P: modify the data collected for frequencies
The previous data collected for each used frequency was not
sufficient/relevant. Remove the counter of the number of interfaces
using the frequency, and track if the frequency is used by a station
interface and/or a P2P client.

Change-Id: I02d9114c4fdc62532911fa0619fd2d8cb217d3f1
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/26343
Tested-by: IWL Jenkins
Reviewed-by: Max Stepanov <Max.Stepanov@intel.com>
Reviewed-by: David Spinadel <david.spinadel@intel.com>

x-iwlwifi-hostap-dev-build: 64b33d3fcf7a7667d0a5a50383bf40071e8a3ce5