wpa_supplicant: Fix iteration over the radio interfaces
On exit from the radio interfaces iteration is done, ifs is always a
valid pointer and thus cannot be used to check if a valid interface
was found or not.

Fix this by testing if the wmm_ie is valid or not.

Change-Id: Ida850c4d115ab0c2a8fe38578fe84d65732b33c8
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/28340
Reviewed-by: David Spinadel <david.spinadel@intel.com>
Tested-by: David Spinadel <david.spinadel@intel.com>

x-iwlwifi-hostap-dev-build: ff40810c886ab51d81beed270f2892b0aa781ade