[WCD] Android: add SETBAND driver command support
When setting a specific band, scan and connection will be done only
on that band. Also, when already connected on the non configured band,
the supplicant will disconnect from that station.
The supplicant will choose the preferred network in the given band and will try
to connect to it, regardless of previous state.

This command is not permitted for p2p device or ap/go interface,
and will be rejected.

Change-Id: Iee59cc6204270830edb3531bd331bb059037fb75
Signed-off-by: Beni Lev <beni.lev@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/28361
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: 27422586af7841dcbb58e5be804c24f9e04f4e4f