[WCD-INTERNAL] nl80211: propagate miracast hint to the driver
When a miracast start/stop indication is received from Android,
propagate this information to the driver using vendor command that
configured the driver to enter/exit low latency mode for the given
interface.

Note that this is a temporary implementation that should be removed
once proper support for low latency is in place.

Change-Id: Id962be1249e358025cb17d99f5ea5a0169da36ca
Signed-off-by: David Spinadel <david.spinadel@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/28898
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: 2191a39a2709cac279f00da11b823506c17a0597