[WCD-INTERNAL] wpa_supplicant: Handle bgscan in case of miracast indication
Disable background scanning in case that Android sends the miracast
indication, enable re-enable it when the miracast indication is removed.

Note that this is a temporary change, and it should be removed once
proper handling for traffic indications and low latency are in place.

Change-Id: I8365ea5b05c4e26580a74d896c1dc30c2e5740e5
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/28931
Reviewed-by: ArikX Nemtsov <arik@wizery.com>
Tested-by: IWL Jenkins
Reviewed-by: Alexander Bondar <alexander.bondar@intel.com>

x-iwlwifi-hostap-dev-build: 9dace84709e0da41c0eb366853b4e67b83686a5e