wpa_supplicant: freq priority check in scan result sort
Added a channel priority field to the scan result struct.
The value is compared in the scan result sorting: if SNR of two APs
is quite close we prefer an AP operating on the channel with a higher
frequency priority.

Change-Id: I0c5261c98e697aa51e8eb011befd0e44edefc677
Signed-off-by: Max Stepanov <Max.Stepanov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/29594
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Tested-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: 13d2fc76a330ccff62bdb4b0c9a54f19e7fc26b0