P2P: move a GO to a frequency that is also supported by the client
A P2P GO interface that was instantiated after a GoN holds the
intersection of frequencies between the GO and the client. In case
that the GO is going to move to anther frequency, allow it to
move only to a frequency that is also supported by the client.

Change-Id: Idcc3796ddc713e9a198f8d1f97b92f8e8d546c50
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/29610
Tested-by: IWL Jenkins
Reviewed-by: David Spinadel <david.spinadel@intel.com>

x-iwlwifi-hostap-dev-build: c0393d489ee610f17dc55cd55b573e0da596060e