[WCD] P2P: add support for using indoor channels
Generally, indoor channels cannot be used for P2P operation as
currently no way to determine that the device is operating in
an indoor environment.

Add a relaxation to allow using indoor channels, iff the peer
device is going to be the GO and in addition it is identified
as a non-mobile/indoor device based on its WPS device category.

Change-Id: I870dab4873627ee063ca6f2e1b291a36c1ef04af
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/29752
Tested-by: IWL Jenkins
Reviewed-by: David Spinadel <david.spinadel@intel.com>

x-iwlwifi-hostap-dev-build: 439125100f7b4ef5f0dbf92aa932ea2d4c928787