[WCD] driver_nl80211: set country on driver COUNTRY command
When a "COUNTRY" driver command arrives on the STA interface, set the
country code to kernel. Set it as a regulatory cell-base hint, as it
is coming from the radio.

Avoid doing so when the P2P-device interface is used, to avoid
double-setting the code. The Android framework always sends the code
down twice.

Convert the special "no-country" code ZZ to the cfg80211 equivalent
(99).

Change-Id: Ia028fd2b864d2e780230295ef11791d1adc5cbff
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/33346
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: c608197284f0c9a7bde25c60da41ece0c51d6b65