[WCD] query country on channel list update
The country code sent on the way down might be "no-country". In case the
kernel updated to a different default country, query the value when
it indicates the regdomain was updated.

Avoid setting the country code on the way down from the Android
framework for the same reason.

Change-Id: I2786d3461971864d4d7020aace5b244bd41abc46
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/33345
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: 81d7e894a8b7a5d89ec61cc5c603c4ec35ed30f4