P2P: select a random social channel as the operating channel
Currently, when an operating channel is not set in the configuration
file a random social channel is selected as the operating channel.
However, during runtime this channel might no longer be valid, e.g.,
when the frequency is disallowed from the control interface.

Modify this so the random selection would be done as part of
best channel selection.

Change-Id: I8e515637f0c419d6471f230d933bc8a7c6491d43
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/31013
Reviewed-by: David Spinadel <david.spinadel@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-hostap-dev-build: e3610961e82a7090c538d9b620a7cb998e779f7f