P2P: Add all channels to GoN response
The P2P specification states that in case the P2P Device
is to become a P2P Client it can set the channels in the
channel list attribute independently from the channel list
attribute in the GoN request.

Add all the P2P Device supported channels to the GoN response,
as this can be useful for cases that the allowed channels of the
peer P2P Device change, and the GO can now operate in additional
channels.

Change-Id: Ifc8788f6a7c72ea6a7681041f9cfcd009b229231
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/31231
Reviewed-by: David Spinadel <david.spinadel@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-hostap-dev-build: 39ee7c65fd4ecfd3932597e723544643aac4d117