P2P: Disallow legacy client connection on indoor channel
In order to prevent daisy chain scenario, a GO should not allow
legacy clients to connect to it.

Add the support to determine if the given operating frequency of
the GO is associated with a channel for which only indoor operation
is allowed and disallow legacy client connection in such cases.

Change-Id: I47300f767a90eb12ea6564c0522fa8848055084b
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/31902
Tested-by: IWL Jenkins

x-iwlwifi-hostap-dev-build: f79dca04b081350283ed424ac3f5f1437704a603