wpa_supplicant: remove pending sme-connect radio work [BUGFIX]
If a new connection is attempted while there is a pending
sme-connection radio work, cancel the pending radio work
and continue with the new connection attempt.

This is preferable over rejecting the new working and
continuing with the pending one, as it is possible that the
previous work is no longer valid.

---

MWG100201156

Change-Id: I3979b2695fe47e1f1b643ff6231230301bba2fd0
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/33544
Tested-by: IWL Jenkins
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/33565

x-iwlwifi-hostap-dev-build: 94e142bdbf8495c7880887c1faf428ecdb213680