P2P: fix usage of group common frequencies [BUGFIX]
The code assumed that the go_params where saved during the life time
of the GO which was a wrong assumption as the go_params were freed
immediately after the interface was configured, and thus the freq_list
that was used for channel selection purposes was invalid.

Fix this by saving the group common frequencies during the lifetime of
the interface.

---

MWG100201166

Change-Id: Ifed9fc9d027692647f8572218a003298989b150c
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/33573
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/33577

x-iwlwifi-hostap-dev-build: 15d5d17031aa00c5aceb57c6f409edc8b621d0c2