P2P: Disallow GO CS immediately after GoN or invitation
A newly created GO might move to another channel before the
client was able to connect to it. This creates a situation
where the client searches the GO on the channel agreed upon
during GON or invitation signaling, while the GO is on another
channel. This in turn might lead to a connection failure and
group removal.

Fix this by disallowing a GO CS as long as there is some activity
that should delay the switch. If a GO move is not allowed, set
a timeout to re-attempt the move.

Change-Id: I1fd4d9f105e5328be8631c4b3dc506ba8f498075
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34583
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34799
Tested-by: IWL Jenkins

x-iwlwifi-hostap-dev-build: a8054cd959a7c94c46a6fc688d63893ace89a9a3