wpa_supplicant: Raise priority of frequency used by p2p under low latency
Raise the priority of the frequency used by P2P interface when
traffic load on this interface is high or when Voice/Video traffic
is present on this interface.
When settings such as security, privacy, noise etc. are equal, the
sorting of bss table will give higher priority to bss who's frequency
is used by P2P interface, which can result in reducing interference
with the Voice/Video or high traffic.

Change-Id: I139cc51ba6b8094cdb7733a39002cab9bc13a68d
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/31571
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/35787

x-iwlwifi-hostap-dev-build: c1f170a136090d66735fa912daa2a7284e973342