scan: reset normal scan counter when a connection succeeds [BUGFIX]
The normal scan counter is used to trigger a few normal scans before a
scheduled scan is issued.  The reason for doing this is that we get
results faster and some drivers support more SSIDs per normal scan
than during sched scan.  After 2 normal scans, we start using
sched_scan.  But the problem is that when a connection succeeds, we do
not reset this counter, so we will keep using sched scans.

To fix this, now we reset the counter when we switch to WPA_COMPLETED
state.

Change-Id: I28a7df11d54180b8478d70bb3b1e959c2d513979
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---

MWG100201146

I'm not sure we should do this in the WPA_CONNECTED state.  Actually
it might make more sense to do it when entering WPA_AUTHENTICATING or
WPA_DISCONNECTED state.

Additionally, I don't know what changed between LinuxCore4 and
LinuxCore5, because this bug didn't happen before...

In v2:

   * s/normal_counter/normal_scans/ to fix compilation error (damn, I
     fixed this in the DUT, but forgot to do it in my git repo).

In v3:

   * now *really* fixed (didn't commit the file earlier, doh!);

In v4:

   * fixed commit message: removed mention to the sched_scan interval
     and rephrased a bit (Ilan);

   * s/WPA_CONNECTED/WPA_COMPLETED/ since the latter is the correct
     name of the state. ;)

In v5:

   * removed the last sentence in the first paragraph, since it was
     not clear and not very useful anyway;

Change-Id: I28a7df11d54180b8478d70bb3b1e959c2d513979
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34688
Reviewed-by: Alexander Bondar <alexander.bondar@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Tested-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/34849
Reviewed-by: Assaf Azulay <assaf.azulay@intel.com>

x-iwlwifi-hostap-dev-build: c5b9e17052a10ea4e4583278dbc9108ff432e24a