[WCD] nl80211: Send driver state HANGED on station interface
Driver state event "HANGED" might be sent on any of the interfaces.
However, the android handling of the event is done only the context
of the station interface state machine.

Modify the current state, so the event would always be sent on the non
dynamic interface (which should be the station interface)

Change-Id: Ib0f95c9539d96c29696ee05d176f235646016922
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38077
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/38100

x-iwlwifi-hostap-dev-build: 8288729e7fe20de28f3b683d26fdb937c238171d