Check current_ssid on unexpected association event
BZ: 505

This is mainly to keep static analyzers silent since it does not look
like this code path can be reached in practice due to the way
association events are handled and current_ssid is either set before
resched here or the association is rejected. Anyway, if this could be
reached, the wpa_supplicant_set_wpa_none_key() call would end up
dereferencing a NULL pointer, so add an explicit check to make sure that
does not happen.

Change-Id: Ib06169c60ca7f6ef25b067593645bfbd8de166e4
Signed-off-by: Jouni Malinen <j@w1.fi>
Reviewed-on: https://gerrit.rds.intel.com/r/49222
Tested-by: IWL Jenkins
Reviewed-by: Guy Cohen <guy.cohen@intel.com>
Reviewed-by: Ilan Peer <ilan.peer@intel.com>

x-iwlwifi-hostap-dev-build: bf16332e46078d658e3cc045dbdcdba6cdc84852