[WCD] P2P: use correct WMM parameters for GO
The WMM parameters are in the order "BE, BK, VI, VO" (note BE/BK!)
but the current code just moves them over into the queue parameters.
These are in the order "VO, VI, BE, BK", so that clearly isn't the
right thing to do. Fix this by explicitly differentiating the two.

In addition, the settings of a queue's cwmin and cwmax was wrong. The
calculation simply needs to translate the acwmin and acwmax values
using the following:

cwmin = 2 ^ acwmin - 1
cwmax = 2 ^ acwmax - 1

Change-Id: Ia414e2f028fd5681b16d3bd755c5502e5e4144b3
Fixes: 763a7b6428561 ("P2P: Use managed interface WMM parameters for local GO") [internal commit ID]
Reported-by: Amir Parhi <amir.parhi@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43160
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Tested-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43737
Reviewed-by: Guy Cohen <guy.cohen@intel.com>

x-iwlwifi-hostap-dev-build: c5428bd6f1cb5727aa24be0c0d84d9b35bb9cb38