wpa_supplicant: Cancel all scans when STA_AUTOCONNECT=0 is received
When STA_AUTOCONNECT=0 hint is received and wpa_supplicant is not
connected to any BSS, cancel scheduled scan and other scans and mark
this interface as disconnected to prevent it from initiating more
scans and trying to connect. Restart scans when this hint is cleared.

Change-Id: I55b44a4d4898d2fd3d3e83f4328fda0593545fbd
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/48174
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Guy Cohen <guy.cohen@intel.com>

x-iwlwifi-hostap-dev-build: dff7928b749d041f36ed7fc1a4695656ce69a428